requirements.txt
alabaster==0.7.12
Babel==2.6.0
cassandra-driver==3.15.1
certifi==2018.10.15
chardet==3.0.4
cx-Oracle==7.0.0
docutils==0.14
idna==2.7
imagesize==1.1.0
Jinja2==2.10
MarkupSafe==1.0
neo4j==1.7.0rc2
neobolt==1.7.0
neotime==1.7.1
numpy==1.15.2
packaging==18.0
pandas==0.23.4
pyasn1==0.4.4
pyexasol==0.4.2
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
splunk-sdk==1.6.5
urllib3==1.23
websocket-client==0.53.0
